first try at reference completions
